new colors added
