+++ index.html
@@ -1,25 +1,26 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/reaper-scythe.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>GrimReaper</title>
-    <script
-      type="text/javascript"
-      src="https://wotlk.ultimowow.com/static/widgets/power.js"
-    ></script>
-    <script>
-      var aowow_tooltips = {
-        colorlinks: true,
-        iconizelinks: true,
-        renamelinks: true
-      }
-    </script>
-  </head>
 
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
-  </body>
+<head>
+  <meta charset="UTF-8" />
+  <link rel="icon" type="image/svg+xml" href="/reaper-scythe.svg" />
+  <link href="https://fonts.googleapis.com/css2?family=Cinzel:wght@400;600&family=Roboto:wght@300;400;500&display=swap"
+    rel="stylesheet">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>GrimReaper</title>
+  <script type="text/javascript" src="https://wotlk.ultimowow.com/static/widgets/power.js"></script>
+  <script>
+    var aowow_tooltips = {
+      colorlinks: true,
+      iconizelinks: true,
+      renamelinks: true
+    }
+  </script>
+</head>
+
+<body>
+  <div id="root"></div>
+  <script type="module" src="/src/main.jsx"></script>
+</body>
+
 </html>
