+++ src/components/AddDKP/AddDkp.css
@@ -80,7 +80,9 @@
   height: 16px;
   border-radius: 50%;
   background-color: #fff;
-  box-shadow: 32px 0 #fff, -32px 0 #fff;
+  box-shadow:
+    32px 0 #fff,
+    -32px 0 #fff;
   position: relative;
   animation: flash 0.5s ease-out infinite alternate;
 }
@@ -88,15 +90,21 @@
 @keyframes flash {
   0% {
     background-color: #fff2;
-    box-shadow: 32px 0 #fff2, -32px 0 #fff;
+    box-shadow:
+      32px 0 #fff2,
+      -32px 0 #fff;
   }
   50% {
     background-color: #fff;
-    box-shadow: 32px 0 #fff2, -32px 0 #fff2;
+    box-shadow:
+      32px 0 #fff2,
+      -32px 0 #fff2;
   }
   100% {
     background-color: #fff2;
-    box-shadow: 32px 0 #fff, -32px 0 #fff2;
+    box-shadow:
+      32px 0 #fff,
+      -32px 0 #fff2;
   }
 }
