+++ src/components/search/SearchPlayer.css
@@ -1,179 +1,247 @@
 .searchplayer-form-container {
   width: 100%;
-  /* display: flex; */
-  /* justify-content: space-between; */
   position: relative;
+  margin-bottom: 20px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+  gap: 15px;
 }
 
 .searchplayer-form {
-  height: 45px;
-  /* width: 400px; */
-  /* position: relative; */
-
   display: flex;
-  justify-content: center;
   align-items: center;
+  background: rgba(10, 12, 14, 0.8);
+  border: 1px solid #c89b3c;
+  border-radius: 30px;
+  overflow: hidden;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.5);
+  transition: box-shadow 0.3s ease;
+}
+
+.searchplayer-form:focus-within {
+  box-shadow: 0 0 0 3px rgba(200, 155, 60, 0.3);
 }
 
 .searchplayer-form input {
-  height: 30px;
-  padding: 0px 10px;
-  margin: 0px;
-  border: 0px;
-  width: 300px;
-  border-radius: 5px 0px 0px 5px;
+  height: 40px;
+  padding: 0 15px;
+  border: none;
+  background: transparent;
+  color: #f0e6d2;
+  font-size: 16px;
+  width: 250px;
+  outline: none;
 }
 
-.searchplayer-form input:focus {
-  outline: 0px;
-  background-color: #79030391;
+.searchplayer-form input::placeholder {
+  color: #7e7f83;
 }
 
 .searchplayer-form button {
-  border: 0px;
-  margin: 0px;
-  height: 30px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-radius: 0px 5px 5px 0px;
-}
-
-.searchplayer-form button:focus {
-  outline: 0px solid #790303;
-}
-
-.searchplayer-form button:active {
-  outline: 1px solid #790303;
+  height: 40px;
+  padding: 0 20px;
+  border: none;
+  background: linear-gradient(145deg, #c89b3c, #a67c00);
+  color: #0a0c0e;
+  font-weight: bold;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  border-left: 1px solid #c89b3c;
+  font-family: 'Cinzel', serif;
 }
 
 .searchplayer-form button:hover {
-  background-color: #79030331;
+  background: linear-gradient(145deg, #d4af37, #b8860b);
+  transform: scale(1.02);
+}
+
+.searchplayer-form button:active {
+  transform: scale(0.98);
 }
 
 .searchplayer-found {
   position: absolute;
-  width: 320px;
-  background: #120931f5;
-  border-radius: 0px 0px 5px 5px;
-  padding-bottom: 5px;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
+  top: 45px;
+  left: 0;
+  width: 300px;
+  background: #1e1f25;
+  border: 1px solid #c89b3c;
+  border-radius: 8px;
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.5);
+  z-index: 100;
+  overflow: hidden;
+  animation: fadeInDown 0.3s ease;
 }
 
-.searchplayer-form h1 {
-  font-size: 16px;
-  height: 27px;
-  width: 100%;
-  cursor: pointer;
+@keyframes fadeInDown {
+  from {
+    opacity: 0;
+    transform: translateY(-10px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
 
-  display: flex;
-  align-items: center;
-  justify-content: center;
+.searchplayer-found h1 {
+  font-size: 16px;
+  padding: 10px 15px;
+  margin: 0;
+  color: #f0e6d2;
+  cursor: pointer;
+  transition:
+    background 0.3s ease,
+    color 0.3s ease;
+  font-family: 'Roboto', sans-serif;
+  border-bottom: 1px solid rgba(200, 155, 60, 0.2);
+}
+
+.searchplayer-found h1:last-child {
+  border-bottom: none;
+}
+
+.searchplayer-found h1:hover {
+  background: #c89b3c;
+  color: #0a0c0e;
 }
 
 .button-add-dkp-container {
-  position: absolute;
-  height: 100%;
-  width: 100%;
-  right: 0px;
-  top: 0px;
-  pointer-events: none;
-
+  position: static;
+  height: auto;
+  width: auto;
+  pointer-events: auto;
   display: flex;
-  align-items: center;
-  justify-content: flex-end;
+  justify-content: center;
 }
 
 .button-add-dkp {
-  height: 30px;
-  width: 90px;
-  background-color: #790303;
-  margin: 0px;
-  padding: 0px 10px;
-  border: 0px;
-  pointer-events: painted;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.button-add-dkp:focus {
-  outline: 0px;
+  height: 40px;
+  padding: 0 20px;
+  background: linear-gradient(145deg, #c89b3c, #8b6910);
+  border: none;
+  border-radius: 20px;
+  color: #0a0c0e;
+  font-weight: bold;
+  font-family: 'Cinzel', serif;
+  cursor: pointer;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.3);
+  transition: all 0.3s ease;
 }
 
 .button-add-dkp:hover {
-  background-color: #c90707;
+  background: linear-gradient(145deg, #d4af37, #a67c00);
+  transform: translateY(-2px);
+  box-shadow: 0 6px 12px rgba(200, 155, 60, 0.4);
 }
 
 .button-add-dkp:active {
-  background-color: #302828;
+  transform: translateY(0);
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
 }
 
 .date-container {
-  position: absolute;
-  height: 100%;
-  width: 100%;
-  left: 0px;
-  top: 0px;
-  pointer-events: none;
-
+  position: static;
+  height: auto;
+  width: auto;
+  pointer-events: auto;
   display: flex;
-  align-items: center;
-  justify-content: flex-start;
+  justify-content: center;
 }
 
 .date {
-  height: 30px;
-  width: 150px;
-  background-color: #79030393;
-  margin: 0px;
-  padding: 0px 10px;
-  border: 1px solid rgb(63, 26, 26);
-  pointer-events: painted;
-  border-radius: 5px;
+  height: 40px;
+  padding: 0 15px;
+  background: #2a2c33;
+  border: 1px solid #c89b3c;
+  border-radius: 20px;
+  color: #f0e6d2;
   font-size: 14px;
   cursor: pointer;
   position: relative;
-
   display: flex;
-  flex-direction: column;
   align-items: center;
   justify-content: center;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.6); /* sombra */
-  transition: box-shadow 0.3s ease;
+  box-shadow: inset 0 2px 4px rgba(0, 0, 0, 0.5);
+  transition: all 0.3s ease;
+  font-family: 'Roboto', sans-serif;
+}
+
+.date:hover {
+  background: #3a3c44;
+  transform: scale(1.02);
 }
 
 .date h1 {
   font-size: 12px;
+  margin: 0;
+  font-weight: normal;
 }
 
 .list-days {
   position: absolute;
-  bottom: -40px;
-  left: 0px;
-  padding: 5px;
-  background: #797979a6;
-  z-index: 1;
-  border-radius: 5px;
-
+  top: 50px;
+  left: 50%;
+  transform: translateX(-50%);
+  background: #1e1f25;
+  border: 1px solid #c89b3c;
+  border-radius: 8px;
+  padding: 10px;
   display: flex;
   gap: 10px;
+  z-index: 200;
+  animation: fadeIn 0.2s ease;
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
 }
 
 .list-days button {
-  width: 100px;
-  height: 30px;
-  background: #790303;
+  padding: 8px 12px;
+  background: linear-gradient(145deg, #c89b3c, #8b6910);
+  border: none;
   border-radius: 5px;
-  padding: 0px;
-  margin: 0px;
+  color: #0a0c0e;
+  font-weight: bold;
+  cursor: pointer;
+  transition: all 0.2s ease;
+  font-family: 'Cinzel', serif;
+}
 
-  display: flex;
-  justify-content: center;
-  align-items: center;
+.list-days button:hover {
+  background: linear-gradient(145deg, #d4af37, #a67c00);
+  transform: scale(1.05);
+}
+
+@media (max-width: 600px) {
+  .searchplayer-form-container {
+    flex-direction: column;
+    align-items: stretch;
+  }
+  .searchplayer-form {
+    width: 100%;
+  }
+  .searchplayer-form input {
+    width: 100%;
+  }
+  .button-add-dkp-container,
+  .date-container {
+    width: 100%;
+  }
+  .button-add-dkp,
+  .date {
+    width: 100%;
+    justify-content: center;
+  }
+  .searchplayer-found {
+    width: 100%;
+  }
 }
