+++ src/views/Home/Home.css
@@ -1,7 +1,18 @@
 .Home {
   width: 100%;
-  max-width: 1100px;
-  padding: 20px 100px;
-
+  max-width: 1400px;
+  margin: 0 auto;
+  padding: 20px 5%;
   position: relative;
+  /* background-color: #0a0c0e; */
+  /* background-image: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100" opacity="0.1"><path d="M10 10 L90 10 L90 90 L10 90 Z" fill="none" stroke="%23C89B3C" stroke-width="0.5"/><path d="M20 20 L80 20 L80 80 L20 80 Z" fill="none" stroke="%23C89B3C" stroke-width="0.3"/></svg>'); */
+  /* background-repeat: repeat; */
+  /* background-size: 100px 100px; */
+  /* min-height: 100vh; */
+}
+
+@media (max-width: 600px) {
+  .Home {
+    padding: 10px;
+  }
 }
