+++ src/views/Home/Home.jsx
@@ -1,19 +1,32 @@
 import { useState } from 'react'
 import AddDkp from '../../components/AddDKP/AddDkp'
 import DkpsTable from './DkpsTable/DkpsTable'
+import PlayerModal from './DkpsTable/RowPlayer/PlayerModal'
 import './Home.css'
-// import { NavLink } from 'react-router-dom'
 
 const Home = ({ showAddDKP }) => {
   const [buttonShowAddDkp, setButtonShowAddDkp] = useState(false)
+  const [selectedPlayer, setSelectedPlayer] = useState(null)
+
+  const handleCloseModal = () => {
+    setSelectedPlayer(null)
+  }
 
   return (
-    <div className='Home' onClick={() => setButtonShowAddDkp(false)}>
+    <div className='Home'>
       <DkpsTable
         showAddDKP={showAddDKP}
         setButtonShowAddDkp={setButtonShowAddDkp}
+        onPlayerSelect={setSelectedPlayer}
       />
       {buttonShowAddDkp && <AddDkp setButtonShowAddDkp={setButtonShowAddDkp} />}
+      {selectedPlayer && (
+        <PlayerModal
+          player={selectedPlayer.player}
+          alters={selectedPlayer.alters}
+          onClose={handleCloseModal}
+        />
+      )}
     </div>
   )
 }
